add v3 compatible with simplelru
v2 had most of the required functions, so this change adds missing ones
to satisfy the simplelru interface.

To do that, RemoveOldest started returning parameters, unlike being void
as before.

Another behaviour change is that Get and Peek now return
the cached value in case it has already expired to be consistent with
the `simplelru` implementation.

Also, GetExpiration is added.